@@ -12,38 +12,38 @@
     "@mui/styles": "5.6.2",
     "@mui/x-data-grid": "5.10.0",
     "@testing-library/jest-dom": "5.16.4",
-    "@types/node": "12.20.50",
-    "@types/react": "17.0.44",
-    "@types/react-dom": "17.0.16",
+    "@types/node": "17.0.31",
+    "@types/react": "18.0.8",
+    "@types/react-dom": "18.0.3",
     "@types/react-instantsearch-core": "6.10.8",
     "@types/react-instantsearch-dom": "6.12.3",
     "@types/react-lazy-load-image-component": "1.5.2",
     "algoliasearch": "4.13.0",
-    "axios": "0.21.4",
+    "axios": "0.27.2",
     "date-fns": "2.28.0",
     "exceljs": "4.3.0",
     "firebase": "9.7.0",
-    "i18next": "20.6.1",
+    "i18next": "21.6.16",
     "notistack": "2.0.4",
-    "qrcode.react": "1.0.1",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-hook-form": "7.29.0",
-    "react-i18next": "11.16.7",
+    "qrcode.react": "3.0.1",
+    "react": "18.1.0",
+    "react-dom": "18.1.0",
+    "react-hook-form": "7.30.0",
+    "react-i18next": "11.16.8",
     "react-instantsearch-dom": "6.24.1",
     "react-lazy-load-image-component": "1.5.4",
     "react-number-format": "4.9.3",
-    "react-router": "5.3.1",
-    "react-router-dom": "5.3.1",
+    "react-router": "6.3.0",
+    "react-router-dom": "6.3.0",
     "react-scripts": "5.0.1",
     "typescript": "4.6.4",
     "use-pagination-firestore": "0.6.0",
-    "web-vitals": "1.1.2",
-    "workbox-core": "5.1.4",
-    "workbox-expiration": "5.1.4",
-    "workbox-precaching": "5.1.4",
-    "workbox-routing": "5.1.4",
-    "workbox-strategies": "5.1.4"
+    "web-vitals": "2.1.4",
+    "workbox-core": "6.5.3",
+    "workbox-expiration": "6.5.3",
+    "workbox-precaching": "6.5.3",
+    "workbox-routing": "6.5.3",
+    "workbox-strategies": "6.5.3"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -69,5 +69,7 @@
       "last 1 safari version"
     ]
   },
-  "devDependencies": {}
+  "devDependencies": {
+    "@types/react-router-dom": "^5.3.3"
+  }
 }
